move cleanup to postrm

+++ dev-lisp/sbcl/sbcl-0.8.4-r2.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-lisp/sbcl/sbcl-0.8.4-r2.ebuild,v 1.2 2003/10/12 03:45:54 mkennedy Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-lisp/sbcl/sbcl-0.8.4-r2.ebuild,v 1.3 2003/10/12 04:53:53 mkennedy Exp $
 
 DESCRIPTION="Steel Bank Common Lisp (SBCL) is a Open Source development system for ANSI Common Lisp. It provides an interactive environment including an integrated native compiler, interpreter, and debugger. (And it, and its generated code, can also play nicely with Unix when running noninteractively.)"
 HOMEPAGE="http://sbcl.sourceforge.net/"
@@ -103,7 +103,12 @@ pkg_postinst() {
 
 pkg_prerm() {
 	/usr/sbin/unregister-common-lisp-implementation sbcl
+}
+
+pkg_postrm() {
 	# since we modified the .fasl files in postinst, we must manually
 	# remove them here
-	rm -rf /usr/lib/sbcl
+	if [ ! -x /usr/bin/sbcl ]; then
+		rm -rf /usr/lib/sbcl
+	fi
 }
